pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "litellm>=1.80.7",
    "openai>=2.9.0",
    "pillow>=12.0.0",
    "playwright>=1.56.0",
    "python-dotenv>=1.2.1",
    "smolagents>=1.23.0",
]
